+++ code/python_tooling/dashboard/app.py
@@ -4,13 +4,14 @@ from dash import dcc, html
 
 from database.types import PoseType
 
-# NOTE(Jack): These might look unused but they need to be imported here so that hte callbacks can be registered with the
-# Dash instance. Please see the answer here https://community.plotly.com/t/splitting-callback-definitions-in-multiple-files/10583/2
-import callbacks_data_loading
-import callbacks_extracted_target
-import callbacks_pose_graph
-import callbacks_slider
-import callbacks_statistics
+# NOTE(Jack): These might look unused but they need to be imported here so that the callbacks can be registered with the
+# Dash instance. Please see the answer here
+# https://community.plotly.com/t/splitting-callback-definitions-in-multiple-files/10583/2
+import callbacks.data_loading
+import callbacks.extracted_target
+import callbacks.pose_graph
+import callbacks.slider
+import callbacks.statistics
 
 # TODO(Jack): Use a css style sheet instead of individually specifying the properties everywhere! This does not scale.
